+++ libraries/Adler/Adler16.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: Adler16.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.2.1
 //    DATE: 2022-06-14
 // PURPOSE: Arduino Library for calculating Adler-16 checksum
 //     URL: https://github.com/RobTillaart/Adler
@@ -56,13 +56,14 @@ void Adler16::add(uint8_t value)
 
 //  straightforward going through the array.
 //  small footprint
-void Adler16::add(uint8_t * array, uint16_t length)
+//  returns current Adler value
+uint16_t Adler16::add(uint8_t * array, uint16_t length)
 {
   while (length--)
   {
     add(*array++);
   }
-  return;
+  return (_s2 << 8) | _s1;
 }
 
 
@@ -71,7 +72,8 @@ void Adler16::add(uint8_t * array, uint16_t length)
 //  S2 grows quadratic
 //  only do modulo when S2 reaches halfway uint16_t
 //                      and at the end of the loop.
-void Adler16::addFast(uint8_t * array, uint16_t length)
+//  returns current Adler value
+uint16_t Adler16::addFast(uint8_t * array, uint16_t length)
 {
   _count += length;
   uint16_t s1 = _s1;
@@ -92,12 +94,19 @@ void Adler16::addFast(uint8_t * array, uint16_t length)
   }
   _s1 = s1;
   _s2 = s2;
+  return (s2 << 8) | s1;
 }
 
 
 uint16_t Adler16::getAdler()
 {
   return (_s2 << 8) | _s1;
+};
+
+
+uint16_t Adler16::count()
+{ 
+  return _count;
 };
 
 
@@ -111,15 +120,15 @@ void Adler16::add(char value)
 }
 
 
-void Adler16::add(char * array, uint16_t length)
+uint16_t Adler16::add(char * array, uint16_t length)
 {
-  add((uint8_t *) array, length);
+  return add((uint8_t *) array, length);
 }
 
 
-void Adler16::addFast(char * array, uint16_t length)
+uint16_t Adler16::addFast(char * array, uint16_t length)
 {
-  addFast((uint8_t *) array, length);
+  return addFast((uint8_t *) array, length);
 }