+++ libraries/Adler/Adler32.h
@@ -2,7 +2,7 @@
 //
 //    FILE: Adler.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.2.1
 //    DATE: 2022-01-27
 // PURPOSE: Arduino Library for calculating Adler-32 checksum
 //     URL: https://github.com/RobTillaart/Adler
@@ -12,7 +12,7 @@
 #include "Arduino.h"
 
 
-#define ADLER32_LIB_VERSION        (F("0.2.0"))
+#define ADLER32_LIB_VERSION        (F("0.2.1"))
 
 
 //  largest prime below 2^16
@@ -24,21 +24,25 @@ class Adler32
 public:
   Adler32();
 
-  void begin(uint32_t s1 = 1, uint32_t s2 = 0);
+  void     begin(uint32_t s1 = 1, uint32_t s2 = 0);
 
-  void add(uint8_t value);
-  void add(uint8_t * array, uint16_t length);
+  void     add(uint8_t value);
+  //  returns current Adler value
+  uint32_t add(uint8_t * array, uint16_t length);
   //  trade PROGMEM for speed
-  void addFast(uint8_t * array, uint16_t length);
+  //  returns current Adler value
+  uint32_t addFast(uint8_t * array, uint16_t length);
 
   //  wrappers for strings
-  void add(char value);
-  void add(char * array, uint16_t length);
-  void addFast(char * array, uint16_t length);
+  void     add(char value);
+  //  returns current Adler value
+  uint32_t add(char * array, uint16_t length);
+  //  returns current Adler value
+  uint32_t addFast(char * array, uint16_t length);
 
 
   uint32_t getAdler();
-  uint32_t count() { return _count; };
+  uint32_t count();
 
 private:
   uint32_t _s1;