+++ libraries/Adler/test/unit_test_Adler32.cpp
@@ -74,15 +74,17 @@ unittest(test_ADLER32_lorem)
     ad32.add(lorem[i]);
   }
   assertEqual(3972480156, ad32.getAdler());
+  assertEqual(868, ad32.count());
 
   ad32.begin();
-  ad32.add(lorem, strlen(lorem));
+  assertEqual(3972480156, ad32.add(lorem, strlen(lorem)));
   assertEqual(3972480156, ad32.getAdler());
+  assertEqual(868, ad32.count());
 
   ad32.begin();
-  ad32.addFast(lorem, strlen(lorem));
+  assertEqual(3972480156, ad32.addFast(lorem, strlen(lorem)));
   assertEqual(3972480156, ad32.getAdler());
-
+  assertEqual(868, ad32.count());
 }