+++ libraries/Adler/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.1] - 2022-09-20
+
+- return value added to **add(array)** 
+- return value added to **addFast(array, length)** 
+- update documentation (e.g. timing table).
+- moved count() to .cpp
+
 ## [0.2.0] - 2022-06-15
 
 - breaking change